Names now start with Configuration.
Derived methods now receive two using argument lists, the first is for
the configuration.
