+++ modules/core/shared/src/main/scala-3/io/circe/derivation/Configuration.scala
@@ -2,6 +2,7 @@ package io.circe.derivation
 
 object Configuration {
   val default: Configuration = Configuration()
+  val empty: Configuration = Configuration(transformNames = Predef.identity, useDefaults = false, discriminator = None)
 }
 case class Configuration(
   transformNames: String => String = Predef.identity,
@@ -15,16 +16,4 @@ case class Configuration(
   def withoutDefaults: Configuration = this.copy(useDefaults = false)
   def withDiscriminator(discriminator: String): Configuration = this.copy(discriminator = Some(discriminator))
   def withoutDiscriminator: Configuration = this.copy(discriminator = None)
-}
-
-object EnumConfiguration {
-  val default: EnumConfiguration = EnumConfiguration()
-}
-
-final case class EnumConfiguration(
-  decodeTransformNames: String => String = Predef.identity,
-  encodeTransformNames: String => String = Predef.identity,
-) {
-  def withDecodeTransformNames(f: String => String): EnumConfiguration = this.copy(decodeTransformNames = f)
-  def withEncodeTransformNames(f: String => String): EnumConfiguration = this.copy(encodeTransformNames = f)
 }
