+++ modules/core/shared/src/main/scala-3/io/circe/derivation/ConfiguredEncoder.scala
@@ -1,0 +1,51 @@
+package io.circe.derivation
+
+import scala.deriving.Mirror
+import scala.compiletime.constValue
+import Predef.genericArrayOps
+import io.circe.{Encoder, Json, JsonObject}
+import io.circe.syntax._
+
+trait ConfiguredEncoder[A](using conf: Configuration) extends DerivedInstance[A] with Encoder.AsObject[A] {
+  def elemEncoders: Array[Encoder[_]]
+
+  final def encodeWith(index: Int)(value: Any): (String, Json) =
+    (elemLabels(index), elemEncoders(index).asInstanceOf[Encoder[Any]].apply(value))
+  
+  final def encodedIterable(value: Product): Iterable[(String, Json)] =
+    new Iterable[(String, Json)]:
+      def iterator: Iterator[(String, Json)] =
+        value.productIterator.zipWithIndex.map((value, index) => encodeWith(index)(value))
+  
+  final def encodeProduct(a: A): JsonObject =
+    JsonObject.fromIterable(encodedIterable(a.asInstanceOf[Product]))
+  
+  final def encodeSum(index: Int, a: A): JsonObject = encodeWith(index)(a) match {
+    case (k, v) => conf.discriminator match {
+      case None => JsonObject.singleton(k, v)
+      case Some(discriminator) => v.asObject.getOrElse(JsonObject.empty).add(discriminator, Json.fromString(k))
+    }
+  }
+}
+object ConfiguredEncoder {
+  inline final def derived[A](using conf: Configuration = Configuration.default)(using A: Mirror.Of[A]): ConfiguredEncoder[A] =
+    new ConfiguredEncoder[A] with DerivedInstance[A](
+      constValue[A.MirroredLabel],
+      summonLabels[A.MirroredElemLabels].map(conf.transformNames).toArray,
+    ) {
+      lazy val elemEncoders: Array[Encoder[_]] = summonEncoders[A.MirroredElemTypes].toArray
+      
+      final def encodeObject(a: A): JsonObject = inline A match {
+        case m: Mirror.ProductOf[A] => encodeProduct(a)
+        case m: Mirror.SumOf[A] => encodeSum(m.ordinal(a), a)
+      }
+    }
+  
+  inline final def derive[A: Mirror.Of](
+    transformNames: String => String = Configuration.default.transformNames,
+    discriminator: Option[String] = Configuration.default.discriminator,
+  ): ConfiguredEncoder[A] = {
+    given Configuration = Configuration(transformNames, useDefaults = false, discriminator)
+    derived[A]
+  }
+}
