+++ modules/core/shared/src/main/scala-3/io/circe/derivation/ConfiguredEnumDecoder.scala
@@ -1,0 +1,29 @@
+package io.circe.derivation
+
+import scala.deriving.Mirror
+import scala.compiletime.constValue
+import Predef.genericArrayOps
+import io.circe.{Decoder, DecodingFailure, HCursor}
+import io.circe.syntax._
+
+trait ConfiguredEnumDecoder[A] extends Decoder[A]
+object ConfiguredEnumDecoder {
+  inline final def derived[A](using conf: EnumConfiguration = EnumConfiguration.default)(using m: Mirror.SumOf[A]): ConfiguredEnumDecoder[A] = {
+    val name = constValue[m.MirroredLabel]
+    val cases = summonSingletonCases[m.MirroredElemTypes, A](name)
+    val labels = summonLabels[m.MirroredElemLabels].toArray.map(conf.decodeTransformNames)
+    new ConfiguredEnumDecoder[A] {
+      def apply(c: HCursor): Decoder.Result[A] =
+        c.as[String].flatMap { s =>
+          labels.indexOf(s) match
+            case -1 => Left(DecodingFailure(s"enum $name does not contain case: $s", c.history))
+            case index => Right(cases(index))
+        }
+    }
+  }
+
+  inline def derive[R: Mirror.SumOf](decodeTransformNames: String => String = EnumConfiguration.default.decodeTransformNames): Decoder[R] = {
+    given EnumConfiguration = EnumConfiguration.default.withDecodeTransformNames(decodeTransformNames)
+    derived[R]
+  }
+}
