+++ modules/core/shared/src/main/scala-3/io/circe/derivation/DerivedInstance.scala
@@ -1,0 +1,6 @@
+package io.circe.derivation
+
+trait DerivedInstance[A](
+  final val name: String,
+  final val elemLabels: Array[String]
+)
